services/product/pyproject.toml
[project]
name = "product-service"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "sqlalchemy==2.0.40",
    "alembic==1.15.2",
    "strawberry-graphql[fastapi]==0.262.5",
    "pydantic-settings==2.9.1",
    "dependency-injector==4.46.0",
    "psycopg2-binary==2.9.10",
]

[dependency-groups]
dev = [
    "pytest==8.3.5",
    "httpx==0.28.1",
]
